requirements.txt
asgiref==3.8.1
Django==5.2.1
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
mysqlclient==2.2.7
sqlparse==0.5.3
typing_extensions==4.13.2
